// case_study/furama_resort/casestudy/src/app/model/facility.ts
export interface Facility {
  id?: number;
  name?: string;
  area?: number;
  cost?: number;
  maxPeople?: number;
  rentType?: string;
  standardRoom?: string;
  descriptionOfOtherConvenience?: string;
  poolArea?: number;
  numberOfFloors?: number;
  attachFacility?: string;
  facilityFree?:string;
  facilityType?:string;
}